replace paillier impl
