+++ python/fate/arch/tensor/phe/__init__.py
@@ -1,0 +1,5 @@
+from ._keypair import PHETensorCipher
+from ._ops import *
+from ._tensor import PHETensor
+
+__all__ = ["PHETensor", "PHETensorCipher"]
