--- python/fate/arch/tensor/phe_vertor_based/__init__.py
@@ -1,5 +1,0 @@
-from ._keypair import PaillierTensorCipher
-from ._ops import *
-from ._tensor import PaillierTensor
-
-__all__ = ["PaillierTensor", "PaillierTensorCipher"]
